anti cheat negative dps bug fix.
fixed a bug in which the player would do negative damage over time.
this occured as the room speed had been changed to 60 steps a second
where i had though it was at the default of 30. As such the stored
server time counted up twice as it should have making it possible for
the program to close and reopen with the stored server time being higher
then the recived server time.